Move getBlockedSites to state/reader
