+++ client/me/site-blocks/main.jsx
@@ -16,7 +16,7 @@ import SectionHeader from 'components/section-header';
 import MeSidebarNavigation from 'me/sidebar-navigation';
 import PageViewTracker from 'lib/analytics/page-view-tracker';
 import QuerySiteBlocks from 'components/data/query-site-blocks';
-import getBlockedSites from 'state/selectors/get-blocked-sites';
+import getBlockedSites from 'state/reader/site-blocks/selectors/get-blocked-sites';
 import isFetchingSiteBlocks from 'state/selectors/is-fetching-site-blocks';
 import getSiteBlocksCurrentPage from 'state/selectors/get-site-blocks-current-page';
 import getSiteBlocksLastPage from 'state/selectors/get-site-blocks-last-page';
